+++ odoo/local-src/sf_sales_forecast/__manifest__.py
@@ -1,0 +1,21 @@
+# -*- coding: utf-8 -*-
+# Part of senseFly.
+
+{
+    'name': 'Sales Forecast',
+    'summary': 'Sensefly dealers sales forecast',
+    'version': '1.0',
+    'author': 'Telmo Santos',
+    'category': 'Sales',
+    'depends': ['sale', 'sales_team'],
+    'data': [
+        'views/sales_forecast.xml',
+        'views/res_config_view.xml',
+        'data/sequence.xml'
+    ],
+    'demo': [],
+    'images': ['static/description/icon.png'],
+    'installable': True,
+    'application': True,
+    'auto_install': False,
+}