+++ odoo/local-src/sf_sales_forecast/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+# Part of sensefly.
+from . import sales_forecast
+from . import res_config